Combined the two into one class cause having two is wasteful
